requirements.txt
absl-py==0.15.0
addict==2.4.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autograd==1.4
autograd-gamma==0.5.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2020.12.5
cffi==1.15.0
cftime==1.5.2
chardet==4.0.0
charset-normalizer==2.0.10
clang==5.0
contextvars==2.4
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.3
flatbuffers==1.12
formulaic==0.2.4
frozenlist==1.2.0
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.48.2
h5py==3.1.0
idna==2.10
idna-ssl==1.1.0
imageio==2.14.1
immutables==0.16
importlib-metadata==4.8.3
interface-meta==1.2.5
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-packaging==0.10.6
jupyter-server==1.13.1
jupyterlab==3.2.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lifelines==0.26.4
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
meshio==4.4.6
mistune==0.8.4
multidict==5.2.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
netCDF4==1.5.8
networkx==2.5
notebook==6.4.7
numpy==1.19.5
oauthlib==3.2.1
open3d==0.14.1
opencv-python==4.1.1.26
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkg_resources==0.0.0
plotly==5.10.0
plyfile==0.7.3
prettytable==2.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.6
psbody-mesh==0.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
pymeshlab==2021.10
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pyvista==0.33.2
PyYAML==5.4.1
pyzmq==22.0.3
rdflib==5.0.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
Rtree==0.9.7
scikit-learn==0.21.3
scipy==1.3.1
scooby==0.5.11
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sniffio==1.2.0
tenacity==8.1.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.1.0
tomlkit==0.8.0
torch==1.3.0
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.3.2
torch-sparse==0.4.3
tornado==6.1
tqdm==4.58.0
traitlets==4.3.3
trimesh==3.9.42
typing-extensions==3.7.4.3
urllib3==1.26.3
vtk==9.1.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wrapt==1.12.1
wslink==1.3.1
yarl==1.7.2
zipp==3.6.0
